refactor: 🧹 remove test file for from_florence_2 function and merge into main test_vlm file
